Add bounce mode to gpio.
This introduces an gpio that triggers once and then goes back to it's basic
state.

Variations are repeat, where the triggering is continuous (creating a
flashlight and autotrigger, where the bounce gpio is triggered on boot
automatically (set to on/off, trigger bounce or start flashing,
dependent on the exact configuration).