Add WLAN bootstrap mode.
This mode replaces the step where you have to flash an "AT-style"
firmware to setup the WLAN. With the SDK 1.2.0 I didn't get it working
anyway that way.

This is how it works, connect your terminal emulator and set it 115200
(or whatever speed you configured earlier). Power up the ESP8266 and
wait for the message to input the ssid and passwd. You have 30 seconds
to do that. Enter ssid, then a space, then the passwd and finally a
linefeed (no carriage return, use ^J instead of enter, if necessary!).

You will be informed of what was received and they're written to the
config. If you keep watching, you will connect progress.