Sensors: init overhaul, start.
Give all sensors an "init" function that also probes the device. If the
device can't be reliably probed, mark the device as such and do no
longer access it (unless explicitly requested by the user with an isr
command). This to prevent multiple devices with the same address being
talked to as the wrong device which may upset the device.

Also move all "static" configuration of the device to the init function.