gpio: add gpio16
This is actually the rtc gpio, meant to be used to wake the esp8266
from deep sleep, at a certain "real time".

This means it's not a real gpio, it's not in the 16 bit bitmap (0-15...)
and it has limited functionality.

On some boards it's wired to the ch_pd or reset pin, in which case you
can't use it.

Available modes:

 - input
 - output
 - bounce/timer

Unavailable modes:

 - counter (needs pin change interrupt)
 - pwm (needs SDK support)
 - i2c

i2c needs an open drain output mode, while gpio16 has, as far as I know,
only a push-pull output mode. If someone can confirm the output can be
set to open drain, I can add the i2c mode.