Add option to disable wlan config bootstrap mode.
Now the config is preserved most of the time, I think this
can be interesting for some users.

If the major config version changes or if the config magic
is wrong, the wlan configuration is still flushed and the
bootstrap mode is entered nevertheless. You will need to
disable it again, manually, after boot.