Create framework for displays. Added saa1064.
The saa1064 is a 4x7 led display multiplexer, controlled over i2c.

Input multiple "slots" of messages that will be shown one at a time,
set a timeout on a message and it will be come automatically after that
time. If no messages are left, show the current time (rtc).

Use 0 for timeout to no auto-expire messages.

All displays (currently saa1064 only) have 8 slots for messages that
can be set using the ds (display-set) command:
    ds <display_id> <slot> <timeout> <message>

Use dd (display-dump) to show all detected displays, add a verbosity
value (0-2) to see more detail.

Finally the db (display-bright) command controls the brightness of the
display. Valid values are 0 (off), 1, 2, 3, 4 (max). Use:
    db <display_id> <brightness>

Please note the saa1064 runs at 5V or higher. It cannot be connected to
the esp8266 directly. It must have it's own 5V power supply and may or
may not need i2c level shifters. I am using level shifters, but it may
not be necessary in the end. YMMV.

Needless to say it will only work if i2c is up and running.