Add "wlan status led" option.
It appears the only status possible is "error" c.q.
"disconnected" which makes the led blink quickly.

The led (gpio) can be shared to a limited level, the
gpio must be set to a (plain) output (not pwm or timer),
so unfortunately it cannot be shared with the generic
status led (for now, maybe in the future).

Also it can't be on the RTC gpio (16), that's a
hardware limitation.