Build only one image at a time.
Determined by the size of the flash or the requested
image type one of plain or rboot is built.

Set IMAGE / FLASH_SIZE_MBITS in the Makefile or add them
to the make command line.

IMAGE=plain or IMAGE=rboot
FLASH_SIZE_MBITS=2,4,8,16 or 32

When a FLASH_SIZE_MBITS > 4 is selected or IMAGE=rboot,
an rboot "BIG FLASH" image is built.

Add the flash size and the image type as a #define to the
CFLAGS, so lean images can be generated for small flash.