Add analog input reading (at last...)
This works with oversampling to filter out the unavoidable
noise and to get a virtual (little bit) higher resolution.

In the background the adc input is sampled continuously and
after 256 samples, the result is averaged and can be queried
using the "analog-read" command.

The returned value is scaled from 0-1023 to 0-65535. That
doesn't mean you get 16 bit resolution, expect some 10-11
bits of real resolution, so scale down (divide) by the
number of bits you're really going to use.