time: overhaul (experimental)
Introduce four distinct clock source:

- system time (as obtained from system_get_time())
- rtc (as obtained from system_get_rtc_time())
- timer (simple millisecond tick, incremented every 100ms from a system timer)
- ntp

The first three time sources count from zero on startup/reset
(and if not, they're biased towards zero) and so they can only
count hours, minutes, seconds from startup. Use the "time-set"
command to set an hour-minute-second offset to approximate real
time.

The ntp clock source of course is completely autonomous, it
increments by itself and also resyncs every 10 minutes. Also it
has knowledge or the real time and date, to it can't be set
manually.

When the time is requested, either the ntp or the rtc clock source
is used. Ntp has preference, but if it fails or hasn't synced yet,
the rtc clock source will be used.

The current approach greatly improves clock accuracy when not ntp
clock source is available.

Depending on what source proves to be the most accurate, one of
them may be selected to be the main source, in the future.