pwm: fix issue where static gpio output was overwritten by ISR.
When all duty cycles are static (i.e. 0 or equal to period), the
ISR is stopped and the relevant GPIO's are set statically on or off.

We allow the ISR to finish the current phase though, to prevent
glitches. While the ISR is running the last phase, it could happen
that some GPIO that was set statically, was overwritten by the ISR
and so was turned on and immediately turned off again.

Fixed this by adding a GPIO mask that the ISR should execute when
switching phase sets (at which point the ISR may also decide to
stop because there are no active duties/channels left).