pwm: fix(?) race condition.
When a second change was requested while the first was still waiting to
be processed (in phase) by the ISR, the second change would get lost
permantly.

This is now solved by updating the previous change and while that
happens, stop the pwm ISR. This will probably produce a small glitch,
but no updates get lost.