Add host-size tool "resetserial" (Linux) that reset (USB) serial port.
When first opened, the port goes into "disconnected" state which lowers
both DTR and RTS. This makes the connected ESP8266 go into reset state
and remain that way.

This tool sets "nochup" mode to the serial port, which means that the
modem control lines are only served by the application (esptool) and not
by the operating system when the port is opened or closed. It also
immediately resets DTR and RTS so the ESP8266 can go ahead and run.