new control code for hitachi-style LCD displays.
The LCD control pins are now restricted to be on one io. With the
new io function "set pin mask" all or selected pins of an io can be
controlled in one go. This function for now has only been implemented
on mcp23017.

It means that all control lines of the LCD connected to a mcp23017 can
be set at once, it's no longer required to that one by one. This means
a huge speedup, much less traffic and delays over the I2C bus.

This should solve the instability at lower I2C clock speeds when using
an LCD connected to a mcp23017.