util: fix bug that caused random watchdog resets.
The logic for comparing to the 32 bit counter (wrap around...) was
wrong, which caused the condition sometimes to be met only after a
complete counter wraparound (26 seconds...).

This trick is to treat both start and current timer values "just" as
unsigned 32 bit integer values and then they will wrap around in
calculations as well, which makes the result correct in the end for
whatever current value of the counter and the target.