i2c-sensors: do sensor initalisation one by one.
Instead of one big chunk. Also now run from the task queue instead of
the timer, which is no longer necessary.

The amount of sensors to be probed using 5 busses and all the i2c bus
recovery time for each failed sensors means the amount of time taken
for this chunk of code may go well over the "sensible" and or watchdog
limits. So do them one by one and yield to the SDK code inbetween.