pwm: add "extended" mode for software based pwm (pwm1).
If set, when the duty period for a pin is set to 1, the cycle frequency
is reduced by 4.

This way a fast, flicker-free cycle frequency can be set, e.g. 305 Hz
for width == 14 bits, but for duty is period 1, the frequency is divided
by 4 to 67 Hz resulting in the duty period to be also reduced by 4,
giving a "1 period" that's really short, comparable to 16 bits PWM.