rboot: create own application interface.
Make an interface to rboot into our own application. This creates a
thin, clean interface to rboot without having to rely on constructs and
compiler options used by rboot itself. The drawback is that the
interface needs to be kept in sync with the rboot api, but that's not
really an issue as the api has been stable for years.

This allows us to remove a few kludges specific for rboot compatibility
(e.g. the ".ci" files).

As all changes to rboot specific for this project now have been
incorporated into the projects code itself, switch back to the original
rboot git repo.

Do the same for esptool2, as our changes have been incorporated now as
well.