espflash: fix bug in flash read / receive with chunk size > 1024.
For payloads > netto UDP or TCP payload, it needs to be broken up into
smaller chunks. The chunks are sent sequentially.

Only the first chunk would be received, now we keep receiving until the
expected amount of bytes is received.

We can't rely on the udp terminating empty packet because it can be
disabled and also tcp doesn't use it.