time: use own implementation of SNTP.
Saves both IRAM and DRAM memory, and works better. It keeps trying
to keep in sync, it doesn't do a one-shot try like the version from
Espressif).

Also we don't need the "sntp_localtime" function from Espressif,
because it's just as well in newlib, as "localtime".

I never understood what it had to do with (s)ntp anyway.