display-orbital: add a flag to explicitly enable detection.
This display is notorious hard to auto-detect. It takes a long time to
startup and in the meantime it does not respond to I2C requests.

Make the detection optional using a config flag. This makes startup quicker
and also allows to try harder / longer to detect it when enabled.