dispatch: enable cmd-from-uart when wlan cannot be connected to.
This allows for bootstrapping the wlan ssid and password using a
terminal emulator over the uart.

It does not disable the initial access point bootstrap mode, which still
can be used.

Beware though that this process enables the cmd-from-uart flag; you may
need to turn it off explicitly when not used or uart to network bridging
is being used.