wlan: fix bug where wlan mode was set to recovery mode permanently when it failed to associate.
This means a temporary connection loss (e.g. caused by not responding
DHCP server, access point temporaly off, etc) would set the ESP
permanently in recovery mode, where it needs to configured for wlan
manually. This is not desirable.

Changed the behaviour now that wlan recovery mode is not saved in the
config, so a reset should be enough to get it working again with the
current wlan parameters.