display: allow sharing the flash read buffer with other users.
If the read buffer was used by another user, re-read the contents from
flash memory. If it's required to write the buffer, simply wait until
the buffer is available.

This allows the picture loading to proceed in parallel with other things
using the flash read buffer, i.e. read/writing the config.