dispatch: fix firmware or image uploads making the command interpreter hang.
Workaround for "pre-announced length commands" (like flash-send <length>) not being
received completely (for any reason) and thereby completely locking any further command
processing. This releases the buffer if the command doesn't get completely received within
two seconds. This could be done at tcp close level as well, but it needs to work for udp
as well and this is more robust.