Fix dhcp server in access point mode.
It appears that if

- IP_SOF_BROADCAST
- IP_SOF_BROADCAST_RECV

are #defined, you also need to explicitly enable broadcasting on a
socket before you're allowed to.

The Espressif DHCP server doesn't do that, so it didn't work.

There is no reason for a simple system like ESP8266 to prohibit or even
regulate broadcasting, so simple turn off this control and enable
everything to broadcast.