log facility overhaul: remove logf() and add date/time support.
logf() is obsolete now, the compiler now picks the correct log function
itself: if only one parameter is given, this is assumed to be a literal
string and it's copied to the log verbatim. If more parameters are
given, the first is a format string and it's handed to vsnprintf
together with the other parameters.

Depending on two new flags, a date and or time stand be prepended to log
messages (in the log buffer).