Refactor some SHA library calls.
The ROM supplies some SHA functions, but lacks the "SHA1" function, so
it needs to be emulated by calling SHA1_Init, SHA1_Update and
SHA1_Final. Provide a SHA1 function that does exactly this, simplifies
code. Also provide a function called SHA1_text that returns the digest
in text form, which is often used instead of the bianry form.