Fix issue where upload of files of just 0xff would fail.
This would trigger an optimisation where the flash would not be erased
for sectors that are just 0xff. The assumption for this optimisation
is wrong, because the sector still needs to be erased if other values
besides 0xff are already there. This optimisation only holds if both
the programmed sector is only 0xff and the new to be programmed sector
is only 0xff. In that case they are already detected to be exactly the
same and the programming is skipped altogether.