wlan: implement own scanning+association code.
Much slower, but at least the ap associated to is always the one with
the strongest signal and not just "the first encountered" or the "same
one we already know". Also implement own reassocation procedure,
including a fresh scan, so the esp won't try to associate to the same
access point endlessly (which may have been gone or may have another
SSID (= mac address) in the meantime.