wlan: move/combine client configure from http and cli to wlan, fix for associate.
The SSID/password would go out of sync between sdk flash area and
config variable and the system would keep associating to the old SSID
after a wcc or http equivalent command. Now an instant reconnect is
performed which also updates the sdk config flash area with the new
parameters.