io_mcp: add two configuration flags and fix mcp interrupt issue.
Add an option to skip the polling altogether. Polling is not required
when the interrupt line is hookup up. Polling and interrupt handling
can interfere with each other.

Add an option to select between open drain interrupt output and
push-pull output.

Interrupt issue: the mcp would raise an interrupt by setting the
interrupt line high, the interrupt line would be cleared by reading the
relevant registers, but the esp would not trigger anything on the positive
edge, action would take place on the negative edge so the interrupt would
never be cleared (well, at least, before the polling could be disabled,
it would be cleared by the polling routine).

Changed the interrupt polarity to be negative edge (high -> low) and
then interrupt gets acted on immediately.